Do not pass props when they can be derived from another prop
